% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sync.R
\name{detect_gene_id_type}
\alias{detect_gene_id_type}
\title{Detect Gene ID Type from Seurat Object}
\usage{
detect_gene_id_type(obj, verbose = TRUE)
}
\arguments{
\item{obj}{Seurat object}

\item{verbose}{Logical. Print details}
}
\value{
Character: "symbol", "ensembl_id", "ensembl_id_version", "entrez_id", or "unknown"
}
\description{
Detect Gene ID Type from Seurat Object
}
\keyword{internal}
