% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seurat_obj.R
\name{extract_sample_names_from_filenames}
\alias{extract_sample_names_from_filenames}
\title{Extract Sample Names from File Names}
\usage{
extract_sample_names_from_filenames(file_names)
}
\description{
Extract Sample Names from File Names
}
\keyword{internal}
