% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_metadata.R
\name{integrate_metadata_safe}
\alias{integrate_metadata_safe}
\title{Integrate Metadata Safely}
\usage{
integrate_metadata_safe(obj, metadata, verbose)
}
\arguments{
\item{obj}{Seurat object}

\item{metadata}{Data frame with samples and groups columns}

\item{verbose}{Logical. Print progress}
}
\value{
Updated Seurat object
}
\description{
Integrate Metadata Safely
}
\keyword{internal}
