% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seurat_obj.R
\name{match_samples_direct_files}
\alias{match_samples_direct_files}
\title{Match Samples with Metadata (Type B: Direct Files)}
\usage{
match_samples_direct_files(data_dir, metadata_samples, verbose)
}
\description{
Match Samples with Metadata (Type B: Direct Files)
}
\keyword{internal}
