% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seurat_obj.R
\name{match_samples_single_file}
\alias{match_samples_single_file}
\title{Match Samples with Metadata (Type C: Single File)}
\usage{
match_samples_single_file(data_dir, metadata_samples, verbose)
}
\description{
Match Samples with Metadata (Type C: Single File)
}
\keyword{internal}
