% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seurat_obj.R
\name{match_samples_subfolders}
\alias{match_samples_subfolders}
\title{Match Samples with Metadata (Type A: Subfolders)}
\usage{
match_samples_subfolders(data_dir, metadata_samples, verbose)
}
\description{
Match Samples with Metadata (Type A: Subfolders)
}
\keyword{internal}
