% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_metadata.R
\name{merge_seurat_objects_safe}
\alias{merge_seurat_objects_safe}
\title{Merge Seurat Objects Safely}
\usage{
merge_seurat_objects_safe(seurat_list, verbose)
}
\arguments{
\item{seurat_list}{List of Seurat objects}

\item{verbose}{Logical. Print progress}
}
\value{
Merged Seurat object
}
\description{
Merges multiple Seurat objects with error handling
}
\keyword{internal}
