% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{read_data_from_direct_files}
\alias{read_data_from_direct_files}
\title{Read Data from Direct Files (Type B)}
\usage{
read_data_from_direct_files(
  data_dir,
  sample_names,
  file_mapping,
  min_cells,
  min_features,
  verbose
)
}
\description{
Read Data from Direct Files (Type B)
}
\keyword{internal}
