% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{read_data_from_subfolders}
\alias{read_data_from_subfolders}
\title{Read Data from Subfolders (Type A)}
\usage{
read_data_from_subfolders(
  data_dir,
  sample_names,
  min_cells,
  min_features,
  verbose
)
}
\description{
Read Data from Subfolders (Type A)
}
\keyword{internal}
