% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{read_h5_file_direct}
\alias{read_h5_file_direct}
\title{Read H5 File Direct}
\usage{
read_h5_file_direct(file_path, sample_name, min_cells, min_features)
}
\description{
Read H5 File Direct
}
\keyword{internal}
