% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{read_threebrothers}
\alias{read_threebrothers}
\title{Read Three Brothers Format (MTX)}
\usage{
read_threebrothers(folder_path, sample_name, min_cells, min_features)
}
\description{
Read Three Brothers Format (MTX)
}
\keyword{internal}
