% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{read_threebrothers_from_file}
\alias{read_threebrothers_from_file}
\title{Read Three Brothers from File}
\usage{
read_threebrothers_from_file(
  file_path,
  sample_name,
  original_filename,
  min_cells,
  min_features
)
}
\description{
Read Three Brothers from File
}
\keyword{internal}
