% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{read_with_flexible_detection}
\alias{read_with_flexible_detection}
\title{Read with Flexible Detection}
\usage{
read_with_flexible_detection(data_path)
}
\description{
Read with Flexible Detection
}
\keyword{internal}
