% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sync.R
\name{sync_genes_for_list}
\alias{sync_genes_for_list}
\title{Auto Synchronize Genes for a List of Seurat Objects}
\usage{
sync_genes_for_list(seurat_list, species, verbose)
}
\arguments{
\item{seurat_list}{List of Seurat objects}

\item{species}{Character. Species (e.g., "homo", "mus")}

\item{verbose}{Logical. Print progress}
}
\value{
Updated list of Seurat objects
}
\description{
Automatically detects gene ID type and synchronizes using geneSync
}
\keyword{internal}
